[TST/CLN] Turn decorators __main__ section into tests
